requirements.txt
langchain>=0.0.309
langchain-core>=0.1.4
langchain-community>=0.0.10
aiohttp>=3.8.4
pydantic>=2.0.0
requests>=2.28.2
python-dotenv>=1.0.0
Flask>=2.0.0
Werkzeug>=2.0.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
numpy>=1.21.0
torch>=1.13.1
transformers>=4.30.2 
PyYAML>=6.0 # For config file
requests # If any synchronous requests are needed, or legacy code
# Optional:
arxiv>=1.4.0 # If using arXiv search